Move selectors into their own directory
Previously, selectors lived in the container files they were used in.
Now, they all live in the selectors directory, along with their
corresponding tests.
