+++ src/components/sidebar/browse/SubjectsContainer.js
@@ -1,20 +1,15 @@
 import { connect } from 'react-redux';
 import toJS from 'util/to-js';
 import { changeBrowseLevel, fetchCoursesRequest, selectSubjectInBrowse } from 'actions';
-import { isFetchingSelector } from 'selectors';
+import { subjectsSelector, browseIsFetchingSelector } from 'selectors';
 import Subjects from './Subjects';
 
-export function subjectsSelector(state) {
-  return state
-    .getIn(['browse', 'subjects'])
-    .sortBy(subject => `${subject.get('id')} - ${subject.get('name')}`);
-}
 
 /* istanbul ignore next */
 function mapStateToProps(state) {
   return {
     subjects: subjectsSelector(state),
-    isFetching: isFetchingSelector(state),
+    isFetching: browseIsFetchingSelector(state),
   };
 }
