+++ src/components/sidebar/search/SearchContainer.js
@@ -1,10 +1,8 @@
 import { connect } from 'react-redux';
 import toJS from 'util/to-js';
+import { searchViewSelector } from 'selectors';
 import Search from './Search';
 
-export function searchViewSelector(state) {
-  return state.getIn(['search', 'view']);
-}
 
 /* istanbul ignore next */
 const mapStateToProps = state => ({
