+++ src/components/sidebar/search/SearchSectionSelectionContainer.js
@@ -1,22 +1,15 @@
 import { connect } from 'react-redux';
 import toJS from 'util/to-js';
 import { viewSearch, addSectionFromSearch } from 'actions';
-import { scheduledSectionsSelector } from 'selectors';
+import { currentCourseNameSelector, currentSectionsSelector, sectionsSelector } from 'selectors';
 import SectionSelection from '../common/SectionSelection';
 
-export function currentCourseNameSelector(state) {
-  return state.getIn(['search', 'currentCourseName']);
-}
-
-export function currentSectionsSelector(state) {
-  return state.getIn(['search', 'currentSections']);
-}
 
 /* istanbul ignore next */
 const mapStateToProps = state => ({
   currentCourseName: currentCourseNameSelector(state),
   sections: currentSectionsSelector(state),
-  scheduledSections: scheduledSectionsSelector(state),
+  scheduledSections: sectionsSelector(state),
 });
 
 const mapDispatchToProps = {
